sclang: Refactor QcCallback declaration out of QcWebView
This commit relocates QcCallback out of the QcWebView header. QcCallback is not specific to WebView.

Also removes a dangerous "using" in header file.
